+++ abel/__init__.py
@@ -1,8 +1,3 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 from ._version import __version__
 
 # Temporary tools to warn about deprecated functions and parameters.