+++ abel/tests/test_tools_vmi.py
@@ -1,6 +1,3 @@
-from __future__ import absolute_import
-from __future__ import division
-
 import numpy as np
 from numpy.testing import assert_allclose, assert_equal