+++ doc/tools/approx_gaussian.py
@@ -1,4 +1,3 @@
-from __future__ import division, print_function
 import numpy as np
 from matplotlib import pyplot as plt
 import abel