+++ doc/transform_methods/basex-basis.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, division
-
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm