+++ doc/transform_methods/daun-basis.py
@@ -1,6 +1,3 @@
-# -*- coding: UTF-8 -*-
-from __future__ import print_function, division
-
 import numpy as np
 from scipy.linalg import solve_banded