+++ doc/transform_methods/rbasex-bR.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, division
-
 import matplotlib.pyplot as plt
 
 R = 4