+++ examples/example_hansenlaw_Xe.py
@@ -1,15 +1,7 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 import numpy as np
 import matplotlib.pyplot as plt
 
 import abel
-import scipy.misc
 
 # This example demonstrates Hansen and Law inverse Abel transform
 # of an image obtained using a velocity map imaging (VMI) photoelecton