+++ examples/example_hansenlaw_basex.py
@@ -1,10 +1,3 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 import os.path
 
 import numpy as np
@@ -12,9 +5,7 @@ import matplotlib.pyplot as plt
 
 import abel
 
-import scipy.misc
 from scipy.ndimage import shift, zoom
-import bz2
 
 # This example demonstrates both Hansen and Law inverse Abel transform
 # and basex for an image obtained using a velocity map imaging (VMI)
@@ -38,8 +29,7 @@ output_plot  = 'plot_' + name + '_comparison_HansenLaw.png'
 
 # Load an image file as a numpy array
 print('Loading ' + filename)
-imagefile = bz2.BZ2File(filename)
-im = np.loadtxt(imagefile)
+im = np.loadtxt(filename)
 
 print("scaling image to size 501 reduce the time of the basis set calculation")
 im = zoom(im, 0.4892578125)